+++ src/components/IIBD/FloatIIBD.css
@@ -1,0 +1,3 @@
+.float-iibd {
+  composes: static-iibd from './StaticIIBD.css'
+}
