+++ src/components/IIBD/IIBD.css
@@ -1,6 +1,19 @@
-.wrap {
+.iibd {
   width: 100%;
   height: 100%;
-  background-color: #333;
+  position: relative;
   overflow: hidden;
+
+  aside {
+    width: 300px;
+    height: 60px;
+    text-align: right;
+    position: absolute;
+    top: 60px;
+    right: 60px;
+
+    button {
+      margin-right: 20px;
+    }
+  }
 }
