+++ src/components/IIBD/StaticIIBD.css
@@ -1,0 +1,5 @@
+.static-iibd {
+  width: 100%;
+  height: 100%;
+  background-color: #333;
+}
